Re-organize python files.
